refactor(evm-connector): add aux functions for amount conversions, resolve suggested issues
- Add aux functions with traits IntoEvmAmount and IntoAmount to parse and fromat U256 <-> Decimal (Amount)
- Remove complex calculations for amounts
- Remove inconsistencies from config system
- Resolve suggested issues and implement suggested code improvements

+++ libs/alloy-evm-connector/src/across_bridge.rs
@@ -86,14 +86,13 @@ impl CollateralBridge for AcrossCollateralBridge {
             return Err(eyre::eyre!("Invalid source and destination"));
         }
 
-
         // Get designation details
         let source_designation = self.source.read().unwrap();
         let destination_designation = self.destination.read().unwrap();
 
         // Use direct chain_operations.execute_command() instead of arbiter
         let command = ChainCommand::ExecuteCompleteAcrossDeposit {
-            chain_id: source_designation.get_chain_id() as u32,
+            chain_id: source_designation.get_chain_id(),
             from: source_designation.get_wallet_address(),
             to: destination_designation.get_wallet_address(),
             deposit_amount: amount,
